// src/app/backoffice/admin/interfaces/IPrivateUser-details.interface.ts
// Generated by https://quicktype.io
export interface IData {
  data:   IUserDetails[];
  totals: Totals;
}

export interface IUserDetails {
  id:                number;
  username:          string;
  created:           string;
  totalAmount:       null | string;
  totalTransactions: string;
  lastTransaction:   null | string;
  showDetails?:      boolean;
}

export interface Totals {
  amount: string;
}